Update Examples for Gamer:Bit Extension v0.0.8
MakeCode Editor updated recently which caused the code to break due to a bug was initially ignored by the editor. There was a patch to the bug in the gamer:bit extension which was updated to v0.0.8 so the following examples updated.

The names of the following examples were also adjusted for consistency.